consolidate archive, delete, favorite, unfavorite, readd, and unfavorite into send item
